feat: 날짜 포맷팅 함수 utils로 분리 #24
